refactor: 모달 컴포넌트를 역할에 따라 파일별로 분리
