+++ src/components/common/Modal/ModalContext.ts
@@ -1,0 +1,14 @@
+import { createContext, useContext } from 'react';
+import { ModalContextType } from './types/Modal.types';
+
+const ModalContext = createContext<ModalContextType | undefined>(undefined);
+
+export const useModalContext = () => {
+  const context = useContext(ModalContext);
+  if (!context) {
+    throw new Error('useModalContext must be used within a ModalProvider');
+  }
+  return context;
+};
+
+export default ModalContext;
